Create sortAppointments
You should design, implement and test a function to allow the user to sort all appointments in
appointmentArr based on one of the following attributes; Priority, Date, Subject, Date, or Venue
and display the sorted appointments on diary when an attribute is selected in the dropdown list
associated with by “by” and the “Sort Appointments” button is pressed. The sorting orders are
specified as follows:
•
•
•
•
Priority: from “High” to “Low”;
Date: from present time to future time;
Subject: following the order of corresponding unicode values, e.g., from A to Z then a to z ;
Venue: following the order of corresponding unicode values, e.g., from A to Z then a to z.
You should use string handling techniques to extract the corresponding attribute values from the
appointment records and sort the appointments on the basis of these values. When sorting ap-
pointments based on Date, you should firstly restore the Date object of appointments using the
extracted Date and Start Time values and then compare their time value (accessed via getTime())
for sorting.
The bubble sort algorithm introduced in Workshop 9 should be employed for the design of the
function.