Create tallyAppointments
Dependencies:
assignAppointmentToArray
----------------
nly the days occurred with appointments will display. The tallying
plan and searching for min/max plan discussed in Workshop 9 should be employed in the design of
the function.
You should use the followings as the guideline:
• Find the earliest day and the latest day with appointments by using the min and max plan. You
should first restore the Date object of appointments using the Date value extracted from ap-
pointment records using string handling techniques and then compare their time value (accessed
via getTime()) to find the earliest day (the min plan) and the latest day (the max plan).
• Following the tally plan, create an array and make it ready for tallying. The length of array
should be the number of days sitting between the earliest day and the latest day. For example,
if the earliest day is 23/09/2017 and the latest day is 29/09/2017, the tally array should be
with a length of 7.
• Access to each of the appointments in appointmentArr and restore its Date object as discussed
above. Calculate the number of days counting from the appointment day to the earliest day
and use it for the index of the element in the tally array that you will increment the value.
• To display the tally result as the summary on the diary, access to each of the elements in
the tally array and restore the Date object if the value is greater than 0. The Date object
can be restored by using the time value in milliseconds, which is calculated by the number of
milliseconds per day multiplying the number of days in difference to the earliest day (the index)
and then adding it to the time value of the earliest day. You can then access to the value of
day, month, and year by using getDate(), getMonth(), and getFullYear().